RST-1483: Migrate employment_details Default and Existing Values (#111)
There is an error in the default value for the `employment_details` section of a `Claim`. The default value is currently `"{}"` which is a string of empty hash notation as opposed to an empty hash itself. This PR sets the default value to `{}` from this point onwards and also migrates existing instances of `"{}"` to `{}`.

This then enables an ActiveRecord query to be run which checks if the `employment_details` section of a `Claim` is `.empty?`, which will be used to provide a data extraction for RST-1483.